Move core analysis methods to triples module
